requirements.txt
discord.py
beautifulsoup4
urllib
googletrans
